requirements.txt
#Other packages
numpy==2.1.1
pandas==2.2.3
psycopg2_binary==2.9.9
python-dotenv==1.0.1
setuptools==74.0.0

#local packages
-e .
